+++ src/components/resources.ts
@@ -66,7 +66,7 @@ export class ResourcesComponent {
     
     try {
       try {
-        builder.add_from_file('/usr/share/com.obision.ObisionSystem/ui/resources.ui');
+        builder.add_from_file('/usr/share/com.obision.ObisionAppSystem/ui/resources.ui');
       } catch (e) {
         builder.add_from_file('data/ui/resources.ui');
       }